+++ lib/helpers/string.rb
